chore: add test to template, add new plugin test command
